Fixed: weights type to float

+++ konfuzio_sdk/trainer/file_splitting.py
@@ -318,7 +318,7 @@ class MultimodalFileSplittingModel(AbstractFileSplittingModel):
                 outputs = model(**inputs)
                 logits = outputs.get("logits")
                 # compute custom loss (suppose one has 3 labels with different weights)
-                loss_fct = nn.CrossEntropyLoss(weight=torch.tensor(class_weights, device=model.device))
+                loss_fct = nn.CrossEntropyLoss(weight=torch.tensor(class_weights, device=model.device, dtype=torch.float))
                 loss = loss_fct(logits.view(-1, self.model.config.num_labels), labels.view(-1))
                 return (loss, outputs) if return_outputs else loss
         trainer = CustomTrainer(
